checking file size now

+++ main.py
@@ -10,6 +10,24 @@ dispatcher = updater.dispatcher
 PATH = "./tmpdl/"
 
 
+def downloadVideo(link):
+    yt = YouTube(link)
+    yt = yt.streams.filter(progressive=True, file_extension='mp4').order_by(
+        'resolution').desc().first()
+    if not os.path.exists(PATH):
+        os.makedirs(PATH)
+    yt.download(PATH)
+    filename = PATH + yt.title + ".mp4"
+    filesize = os.path.getsize(filename) / (1024*1024)
+    if filesize > 20:
+        print(f"File {filename} too big with {filesize} MB.")
+        os.remove(filename)
+        return "errtoobig"
+    else:
+        print(f"File {filename} will be sent with {filesize} MB.")
+        return filename
+
+
 def createAnswer(keywords, isQuickplay):
     terms = ' '.join([word for word in keywords])
     results = YoutubeSearch(terms, max_results=1).to_dict()
@@ -24,15 +42,14 @@ def createAnswer(keywords, isQuickplay):
 
 def quickplay(update, context):
     link = createAnswer(context.args, True)
-    yt = YouTube(link)
-    yt = yt.streams.filter(progressive=True, file_extension='mp4').order_by(
-        'resolution').desc().first()
-    if not os.path.exists(PATH):
-        os.makedirs(PATH)
-    yt.download(PATH)
-    context.bot.send_video(
-        chat_id=update.message.chat_id, video=open(PATH + yt.title + ".mp4", "rb"), supports_streaming=True)
-    os.remove(PATH + yt.title + ".mp4")
+    answer = downloadVideo(link)
+    if answer == "errtoobig":
+        context.bot.send_message(
+            chat_id=update.message.chat_id, text="File too big!")
+    else:
+        context.bot.send_video(
+            chat_id=update.message.chat_id, video=open(answer), supports_streaming=True)
+        os.remove(answer)
 
 
 def play(update, context):
@@ -45,18 +62,18 @@ def dl(update, context):
     if(len(context.args[0]) > 1):
         yt = YouTube(context.args[0])
         context.bot.send_message(
-            chat_id=update.message.chat_id, text=f"Downloading {yt.title}")
+            chat_id=update.message.chat_id, text=f"Downloading: {yt.title}")
     else:
         context.bot.send_message(
             chat_id=update.message.chat_id, text="No link found!")
-    yt = yt.streams.filter(progressive=True, file_extension='mp4').order_by(
-        'resolution').desc().first()
-    if not os.path.exists(PATH):
-        os.makedirs(PATH)
-    yt.download(PATH)
-    context.bot.send_video(
-        chat_id=update.message.chat_id, video=open(PATH + yt.title + ".mp4", "rb"), supports_streaming=True)
-    os.remove(PATH + yt.title + ".mp4")
+    answer = downloadVideo(context.args[0])
+    if answer == "errtoobig":
+        context.bot.send_message(
+            chat_id=update.message.chat_id, text="File too big!")
+    else:
+        context.bot.send_video(
+            chat_id=update.message.chat_id, video=open(answer), supports_streaming=True)
+        os.remove(PATH + yt.title + ".mp4")
 
 
 # start polling
